Replace the locking of instances with a modifier of the name instead
The game didn't like that the addon tried to change what instances where being selected/unselected when creating a new group. This created an error which caused the entire UI to break.

I've so far been unable to come up with a way to lock instances without this happening and have thus changed it to instead modify the text by adding a "locked" icon next to the checkbox to give a heads up that the character is saved.